workers/ai_processor/requirements.txt
anthropic==0.18.1
openai==1.12.0
redis==4.6.0
pydantic==2.6.1
sentry-sdk==1.40.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
httpx==0.26.0
